backend/paysignal_django/requirements.txt.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.7.0
djangorestframework==3.15.2
pillow==11.1.0
sqlparse==0.4.4
tzdata==2023.3